requirements.txt
Django==1.6.5
MySQL-python==1.2.5
PIL==1.1.7
South==0.8.4
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.18
boto==2.29.1
celery==3.1.12
django-celery==3.1.10
django-collage==1.0.1
django-common-helpers==0.6.0
django-cron==0.3.3
django-debug-toolbar==1.2.1
django-registration==1.0
kombu==3.0.20
oauthlib==0.6.3
pytz==2014.4
requests==2.1.0
requests-oauthlib==0.4.0
sorl-thumbnail==11.12.1b
sqlparse==0.1.11
twython==3.1.2
wsgiref==0.1.2
